Split out proto.
